requirements.txt
requests==2.31.0
pandas==2.1.3
SQLAlchemy==2.0.23
pymysql==1.1.0
python-dotenv==1.0.1
streamlit==1.33.0
altair


